pyproject.toml
[project]
name = "diffusion-lab"
version = "0.1.0"
description = "Educational lab for learning diffusion models with HuggingFace"
authors = [
    {name = "Diffusion Lab", email = "lab@example.com"}
]
dependencies = [
    "diffusers>=0.21.0",
    "transformers>=4.25.0",
    "accelerate>=0.16.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "Pillow>=9.0.0",
    "matplotlib>=3.5.0",
    "numpy>=1.21.0",
    "jupyter>=1.0.0",
    "ipywidgets>=8.0.0",
    "tqdm>=4.64.0",
    "huggingface-hub>=0.16.0",
    "datasets>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
]